Replace original 'arrowfn.jstest' with ported SpiderMonkey's test cases
Although original tests covered most of the cases, they were, like
implementation itself, based on now outdated spec. So as a first step it
was decided to port more up to date tests from V8 or SpiderMonkey.
(There's no tests for ES6 in the official [TC39 repo] yet.)

And because [V8's test suite] appears to be really small and cover only
basics, it's [SpiderMonkey's test cases] that was selected as the most
complete.

Unfortunately, right now some of the tests fail and were ignored.

[test262 repo]: https://github.com/tc39/test262
[SpiderMonkey's test cases]: https://github.com/mozilla/gecko-dev/tree/master/js/src/jit-test/tests/arrow-functions
[V8's test suite]: https://code.google.com/p/v8/source/browse/trunk/test/mjsunit/harmony/arrow-functions.js